Code for Robocar
This code is for the working of the Robocar as the receiving side. The Robocar has modules like Arduino UNO, which receives the signal through HC-12 wireless Module, then process it and gives the signal to the L298n motor driver for the proper functioning of the robocar according to the given signal.