Code for Gesture Sensor
This code is for the gesture sensor i.e. PAJ7620 which receives the gestures and then sends it to the WEMOS D1 mini microcontroller as a signal for processing the gesture and further sending it to the HC-12 wireless module, to transfer the signal to the robocar to receive it and work accordingly.